fix(echo): break outbound-pipeline import cycle + make channelOutbound backward-compatible (ClawSweeper P1)
Extracted the channel-outbound types (MirrorDispatcher, ChannelEchoAdmission, EchoAdmissionTarget, ChannelOutboundRegistrar) into a leaf module channel-outbound-registrar.types.ts that imports only leaf type modules; ChannelGatewayContext + the runtime modules re-export from it, so the type no longer drags the outbound delivery pipeline (mirror-dispatch -> echo -> deliver -> types.adapters) into a madge cycle. ChannelGatewayContext.channelOutbound is now optional for backward-compat (host always injects it; pre-existing SDK consumers stay valid). Verified: tsgo src+ext 0, check:madge-import-cycles 0, oxlint 0.
