+++ src/channels/plugins/types.adapters.ts
@@ -11,7 +11,7 @@ import type { GroupToolPolicyConfig } from "../../config/types.tools.js";
 import type { ChannelApprovalNativeRuntimeAdapter } from "../../infra/approval-handler-runtime-types.js";
 import type { ChannelApprovalKind } from "../../infra/approval-types.js";
 import type { ExecApprovalRequest, ExecApprovalResolved } from "../../infra/exec-approvals.js";
-import type { ChannelOutboundRegistrar } from "../../infra/outbound/channel-outbound-registrar.js";
+import type { ChannelOutboundRegistrar } from "../../infra/outbound/channel-outbound-registrar.types.js";
 import type {
   PluginApprovalRequest,
   PluginApprovalResolved,
@@ -255,9 +255,11 @@ export type ChannelGatewayContext<ResolvedAccount = unknown> = {
    * Host-issued, owner-bound registrar for the pin-from-here mirror-dispatcher and
    * echo-admission registries. The gateway binds it to THIS channel's authenticated
    * id, so a channel plugin can register/unregister only its own mirror/admission
-   * handlers and cannot spoof or hijack another channel.
+   * handlers and cannot spoof or hijack another channel. Optional in the type for
+   * backward-compatibility (the host always injects it at runtime; pre-existing SDK
+   * channel consumers that predate it still construct a valid context and optional-chain it).
    */
-  channelOutbound: ChannelOutboundRegistrar;
+  channelOutbound?: ChannelOutboundRegistrar;
   /**
    * Optional channel runtime helpers for external channel plugins.
    *
